+++ tests/assertions/assert_url.test.js
@@ -1,11 +1,12 @@
 "use strict";
 
 const assert = require('assert');
-const Wendigo = require('../lib/wendigo');
-const utils = require('./utils');
-const configUrls = require('./config.json').urls;
+const Wendigo = require('../../lib/wendigo');
+const utils = require('../utils');
+const configUrls = require('../config.json').urls;
 
-describe("Not Assertions", function() {
+
+describe("Assert Url", function() {
     this.timeout(5000);
     let browser;
 
@@ -17,35 +18,25 @@ describe("Not Assertions", function() {
         await browser.close();
     });
 
-    it("Not Visible", async () => {
-        assert(browser.assert.not.visible);
+    it("Url", async () => {
         await browser.open(configUrls.index);
-        await browser.assert.not.visible(".hidden-text");
-        await browser.assert.not.visible(".hidden-text2");
+        await browser.assert.url(configUrls.index);
     });
 
-    it("Not Visible Not Exists", async () => {
-        assert(browser.assert.not.visible);
+    it("Url Throws", async () => {
+        const invalidUrl = "http://localhost/not_the_url";
         await browser.open(configUrls.index);
-        await browser.assert.not.visible(".imaginary-text");
+        await utils.assertThrowsAssertionAsync (async () => {
+            await browser.assert.url(invalidUrl);
+        }, `Expected url to be "${invalidUrl}", "${configUrls.index}" found`);
     });
 
-    it("Not Visible Throws", async () => {
-        await browser.open(configUrls.index);
-        assert(browser.assert.not.visible);
-        await utils.assertThrowsAssertionAsync (async () => {
-            await browser.assert.not.visible("p");
-        }, `Expected element "p" to not be visible`);
-        await utils.assertThrowsAssertionAsync (async () => {
-            await browser.assert.not.visible("h1");
-        }, `Expected element "h1" to not be visible`);
-    });
-
-    it("Not Visible Throws With Custom Message", async () => {
+    it("Url Throws With Custom Message", async () => {
+        const invalidUrl = "http://localhost/not_the_url";
         await browser.open(configUrls.index);
         await utils.assertThrowsAssertionAsync (async () => {
-            await browser.assert.not.visible("p", "not visible failed");
-        }, `not visible failed`);
+            await browser.assert.url(invalidUrl, "invalid url");
+        }, `invalid url`);
     });
 
     it("Not Url", async () => {
@@ -69,5 +60,4 @@ describe("Not Assertions", function() {
             await browser.assert.not.url(configUrls.index, "not url failed");
         }, `not url failed`);
     });
-
 });